move interpolation and math module to reverie-util
