+++ reverie-util/src/interpolation/types.rs
@@ -1,0 +1,3 @@
+pub type Time = i64;
+pub type TimeSpan = u32;
+pub(crate) type NormalizedTime = f32;
