+++ reverie/src/lib.rs
@@ -2,7 +2,6 @@ mod context;
 mod engine;
 pub mod gl;
 pub mod gui;
-mod interpolation;
 pub mod math;
 pub mod platform;
 pub mod shader;
@@ -13,4 +12,4 @@ pub mod window;
 pub use context::{Context, ContextBackend};
 pub use engine::ReverieEngine;
 pub use gui::VaoBuilder2DGui;
-pub use interpolation::Interpolation;
+pub use reverie_util as util;
